task: Add robots.txt route
