+++ google_pathways_api/app.py
@@ -3,7 +3,8 @@ from flask_migrate import Migrate
 
 from google_pathways_api.config.config import ConfigurationFactory
 from google_pathways_api.db.models import db
-from google_pathways_api.resources.pathways import blueprint
+from google_pathways_api.resources.pathways import pathways_blueprint
+from google_pathways_api.resources.static_resources import static_blueprint
 
 
 def create_app():
@@ -14,6 +15,7 @@ def create_app():
     db.init_app(app)
     migrate = Migrate(app, db)
 
-    app.register_blueprint(blueprint)
+    app.register_blueprint(pathways_blueprint)
+    app.register_blueprint(static_blueprint)
 
     return app
