+++ google_pathways_api/templates/sitemap.xml
@@ -3,7 +3,7 @@
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
 
    <url>
-      <loc>{{ base_url }}pathways</loc>
+      <loc>{{ base_url }}/pathways</loc>
       <lastmod>2020-04-24</lastmod>
       <changefreq>monthly</changefreq>
       <priority>0.8</priority>
@@ -11,8 +11,8 @@
 
    {% for program in pathways_programs %}
    <url>
-      <loc>{{ base_url }}pathways?page={{loop.index}}</loc>
-      <xhtml:link rel="amphtml" href="{{ base_url }}?page={{loop.index}}" />
+      <loc>{{ base_url }}/pathways?page={{loop.index}}</loc>
+      <xhtml:link rel="amphtml" href="{{ base_url }}/?page={{loop.index}}" />
       <lastmod>{{ program.updated_at }}</lastmod>
       <changefreq>weekly</changefreq>
    </url>
